Use JsonConfiguration for sections instead of a new class
It allows us to create simple Key-Value pairs in the settings.

Example:

    { "user": "Marco" }

is now a valid json configuration.
